% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspirin}
\alias{aspirin}
\title{Aspirin Myocardial Infarction Trial}
\format{
A data frame with 1 row and 4 variables:
\describe{
  \item{y0}{Number of MIs in the placebo group (26).}
  \item{N0}{Number of subjects in the placebo group (11,034).}
  \item{y1}{Number of MIs in the aspirin group (10).}
  \item{N1}{Number of subjects in the aspirin group (11,037).}
}
}
\usage{
aspirin
}
\description{
Data from the Physicians' Health Study (1989), a large-scale
  randomized trial studying the effect of aspirin on myocardial infarction
  (MI). 22,071 male physicians were randomized to aspirin or placebo.
}
\examples{
data(aspirin)
brease(aspirin$y0, aspirin$y1, aspirin$N0, aspirin$N1)
}
\references{
Steering Committee of the Physicians' Health Study Research Group (1989).
Final report on the aspirin component of the ongoing Physicians' Health
Study. \emph{New England Journal of Medicine}, 321(3), 129--135.

Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary
Experiments. \emph{Bayesian Analysis}. \doi{10.1214/25-BA1506}
}
\keyword{datasets}
