% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspirin_meta}
\alias{aspirin_meta}
\title{Aspirin Meta-Analysis Dataset}
\format{
A data frame with 13 rows and 5 variables:
\describe{
  \item{Study}{Study name.}
  \item{y1}{Number of MI events in the aspirin (treatment) group.}
  \item{N1}{Number of subjects in the aspirin (treatment) group.}
  \item{y0}{Number of MI events in the placebo (control) group.}
  \item{N0}{Number of subjects in the placebo (control) group.}
}
}
\usage{
aspirin_meta
}
\description{
Data from 13 randomized controlled trials studying the effect
  of aspirin on myocardial infarction (MI), used for meta-analysis in
  Irons and Cinelli (2025). Each row is a separate study.
}
\examples{
data(aspirin_meta)
head(aspirin_meta)
}
\references{
Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary
Experiments. \emph{Bayesian Analysis}. \doi{10.1214/25-BA1506}
}
\keyword{datasets}
