% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bf_ib_analytic}
\alias{bf_ib_analytic}
\title{Independent Beta analytic Bayes factor (log scale)}
\usage{
bf_ib_analytic(y0, y1, N0, N1, a = 1, b = 1)
}
\description{
Independent Beta analytic Bayes factor (log scale)
}
\keyword{internal}
