% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_count}
\alias{check_count}
\title{Check that a value is a non-negative integer (count)}
\usage{
check_count(x, name)
}
\description{
Check that a value is a non-negative integer (count)
}
\keyword{internal}
