% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_positive_int}
\alias{check_positive_int}
\title{Check that a value is a positive integer}
\usage{
check_positive_int(x, name)
}
\description{
Check that a value is a positive integer
}
\keyword{internal}
