% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_prob}
\alias{check_prob}
\title{Check that a value is a probability (between 0 and 1)}
\usage{
check_prob(x, name)
}
\description{
Check that a value is a probability (between 0 and 1)
}
\keyword{internal}
