% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{lml_brease_mono}
\alias{lml_brease_mono}
\title{Log Marginal Likelihood for the Monotone BREASE Model (H1)}
\usage{
lml_brease_mono(y0, y1, N0, N1, mu0 = 0.5, n0 = 2, mue = 0.5, ne = 2)
}
\arguments{
\item{y0}{Number of events in the control group.}

\item{y1}{Number of events in the treatment group.}

\item{N0}{Number of subjects in the control group.}

\item{N1}{Number of subjects in the treatment group.}

\item{mu0}{Prior mean for baseline risk (default 0.5).}

\item{n0}{Prior sample size for baseline risk (default 2).}

\item{mue}{Prior mean for efficacy (default 0.5).}

\item{ne}{Prior sample size for efficacy (default 2).}
}
\value{
The log marginal likelihood (a scalar).
}
\description{
Computes the log marginal likelihood under the alternative
  hypothesis for the monotone (no-harm) BREASE model, where
  \eqn{\eta_s = 0}.
}
\examples{
lml_brease_mono(y0 = 169, y1 = 9, N0 = 20172, N1 = 19965,
                mu0 = 0.5, n0 = 2, mue = 0.3, ne = 1)

}
\references{
Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary
Experiments. \emph{Bayesian Analysis}. \doi{10.1214/25-BA1506}
}
