% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lml_ib_h1}
\alias{lml_ib_h1}
\title{Independent Beta log marginal likelihood under H1}
\usage{
lml_ib_h1(y0, y1, N0, N1, a0 = 1, b0 = 1, a1 = 1, b1 = 1)
}
\description{
Independent Beta log marginal likelihood under H1
}
\keyword{internal}
