% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{log_prior_brease}
\alias{log_prior_brease}
\title{Log Prior Density for the BREASE Model}
\usage{
log_prior_brease(
  theta0,
  eta_e,
  eta_s,
  mu0 = 0.5,
  n0 = 2,
  mue = 0.5,
  ne = 2,
  mus = 0.5,
  ns = 2
)
}
\arguments{
\item{theta0}{Baseline risk.}

\item{eta_e}{Efficacy.}

\item{eta_s}{Risk of side effects.}

\item{mu0}{Prior mean for baseline risk (default 0.5).}

\item{n0}{Prior sample size for baseline risk (default 2).}

\item{mue}{Prior mean for efficacy (default 0.5).}

\item{ne}{Prior sample size for efficacy (default 2).}

\item{mus}{Prior mean for side-effect risk (default 0.5).}

\item{ns}{Prior sample size for side-effect risk (default 2).}
}
\value{
The log prior density (a scalar).
}
\description{
Evaluates the log prior density of the BREASE model at
  given parameter values.
}
\examples{
log_prior_brease(0.5, 0.3, 0.1)

}
