% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nejm}
\alias{nejm}
\title{NEJM Clinical Trials Dataset}
\format{
A data frame with 39 rows and 10 variables:
\describe{
  \item{No.}{Trial reference number.}
  \item{P.value}{Frequentist p-value.}
  \item{BF01}{Bayes factor in favor of H0.}
  \item{y1}{Number of events in group 1 (control).}
  \item{y2}{Number of events in group 2 (treatment).}
  \item{n1}{Number of subjects in group 1 (control).}
  \item{n2}{Number of subjects in group 2 (treatment).}
  \item{y1n1}{Event rate in group 1 (y1/n1).}
  \item{y2n2}{Event rate in group 2 (y2/n2).}
  \item{ynavg}{Average event rate across both groups.}
}
}
\usage{
nejm
}
\description{
A collection of 39 clinical trials published in the
  \emph{New England Journal of Medicine}, used to compare Bayesian
  analysis methods. Each row is a separate trial with binary outcomes.
}
\examples{
data(nejm)
head(nejm)
}
\references{
Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary
Experiments. \emph{Bayesian Analysis}. \doi{10.1214/25-BA1506}
}
\keyword{datasets}
