% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfizer}
\alias{pfizer}
\title{Pfizer COVID-19 Vaccine Trial}
\format{
A data frame with 1 row and 4 variables:
\describe{
  \item{y0}{Number of COVID-19 cases in the placebo group (169).}
  \item{N0}{Number of subjects in the placebo group (20,172).}
  \item{y1}{Number of COVID-19 cases in the vaccine group (9).}
  \item{N1}{Number of subjects in the vaccine group (19,965).}
}
}
\usage{
pfizer
}
\description{
Data from the Pfizer-BioNTech COVID-19 vaccine trial
  (Polack et al., 2020), a large-scale randomized trial studying vaccine
  efficacy against symptomatic COVID-19. 40,137 participants were
  randomized to vaccine or placebo.
}
\examples{
data(pfizer)
brease(pfizer$y0, pfizer$y1, pfizer$N0, pfizer$N1)
}
\references{
Polack, F.P., Thomas, S.J., Kitchin, N. et al. (2020). Safety and
Efficacy of the BNT162b2 mRNA Covid-19 Vaccine. \emph{New England
Journal of Medicine}, 383(27), 2603--2615.

Irons, N.J. and Cinelli, C. (2025). Causally Sound Priors for Binary
Experiments. \emph{Bayesian Analysis}. \doi{10.1214/25-BA1506}
}
\keyword{datasets}
