% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{sim_ib}
\alias{sim_ib}
\title{Simulate from the Independent Beta Prior}
\usage{
sim_ib(n = 10000, mu0 = 0.5, n0 = 2, mu1 = mu0, n1 = n0)
}
\arguments{
\item{n}{Number of samples to draw.}

\item{mu0}{Prior mean for control group risk (default 0.5).}

\item{n0}{Prior sample size for control group risk (default 2).}

\item{mu1}{Prior mean for treatment group risk (default \code{mu0}).}

\item{n1}{Prior sample size for treatment group risk (default \code{n0}).}
}
\value{
A data.frame with columns: \code{theta0}, \code{theta1}.
}
\description{
Generates prior predictive samples from the independent
  Beta prior on \eqn{(\theta_0, \theta_1)}.
}
\examples{
prior_samps <- sim_ib(10000)
plot(prior_samps$theta0, prior_samps$theta1, pch = ".", col = "gray")

}
